refactor: 채팅 컴포넌트 일부 분리
